requirements.txt
streamlit>=1.30.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
